SMT Guided Counterexample Shrinking
- removes assertions on buffer lengths derived from the max read index
- implements a sound shrinking routines for buffer models

Previous to this PR we inserted constraints for each buffer that stated
that their length could not be larger than the maximum index read from
that buffer. This approach was unsound as branch conditions could
contain arbitrary restrictions on the length of a buffer without reading
from it (via BufLength).

Here we remove these constraints and use the max read index as a
starting point for a counterexample shrinking routine. If our initial
query returns sat, we check if any of the returned buffer models are
very large (we use 1kb as the cutoff), and if they are we run the
initial query again, but restrict the length to the smaller of 1kb and
the max read index. If the query remains sat, we return this model, if
not we double the maximum size and try again.

This should hopefully give us small counterexamples for calldata in a
sound manner.