Solidity: Correct source line in debugger for foundry projects
Solidity sourcemaps refer to files by a numerical id. When building all
source files using a single compiler pass, these identifiers are
sequential in the `sources` key, and hevm previously relied on this
implicitly. When building with multiple parallel compiler passes we can
no longer rely on these identifiers being sequential.

This PR reworks some of the data structures related to source files to
support this potential non sequential source file identification, and
fixes a bug where the wrong line was shown in the solidity pane when
debugging tests from a foundry project.