EVM: respect prank address in callChecks
Previously we would incorrectly fail here if sending value after calling
`vm.prank` since `callChecks` did not take the value of `overrideCaller`
into account when checking if the caller had sufficient balance.

There is some duplication of the logic in `transfer` here, but I'm not
sure how to handle it better, since I do not want `transfer` to be
incomplete, and we can't rely on `transfer` for callcode and
delegatecall (which don't actually move any eth).

This fixes #343.