Workaround for a possible bug in GHC
On arm Mac, the GHC is trying to locate libc++, but during that it also
searches for c++ (without the lib prefix), which is an alias for a clang
executable. It tries to load an executable as a static library which
results in an error.

I have no idea why these changes work, but they seem to work.