Fix bug in SMT encoding of reading multiple consecutive bytes
The current encoding in function `assertReads` asserted, for a read of
more than one (and not 32) bytes, that the same expression is equal to
multiple different constants. This caused some problems to be trivially
unsatisfiable, possibly hiding other bugs.

The new encoding should hopefully fix that.
After running our test suite, it seems like this actually affected some
of our tests, because the overall runtime increased noticeably.