EVM: finalize: fix handling of symbolic returndata
We produced a `Partial` node with a message "runtime code cannot have an
abstract length" if we could not determine a concrete size for
returndata when finalizing a transaction. This changes gates that check
so it is only made if we are actually in a create transaction (which is
the only context where it is valid).