Revert wraparound semantics for reading from buffers
Recently, we have changed the semantics of reading beyond index 2^256
from a buffer. Previously, any read beyond buffer length would return
all zeroes. However, to match the semantics of SMT, we changed that so
that if reading from multiple consecutive indices would overflow the
index (at 2^256), the indices would wrap around and start from 0 again.
Unfortunately, it turns out that the interpreters (such as evmtool) do
not behave this way, they follow the original "just return all zeroes"
strategy.

In order to match the behaviour of real interpreters, we revert the
recent changes and return to the original behaviour.

This also introduces the problem of incompatibility with our SMT
encoding, which should be mitigated by modifying the fuzzer to not
generate these corner cases.