Fuzzer: Generate smaller indices in read expressions
When generating expressions that read multiple consecutive bytes
(ReadWord and CopySlice expressions), we need to avoid generating
indices close to max value of Word256, as these would cause overflow in
indices. Such overflows have different semantics in SMT and in EVM, so
the fuzzer would report unhelpful failures when fuzzing our simplifier.

The proposed solution is to trim concrete Lit expressions in as read
index.