Fix interpretation of buffer model value
Previosuly, we would interpret any buffer whose model is a constant-zero
array as an empty buffer, disregarding the value of the length of the
buffer in the model.
However, the length must be taken into account.

Fun fact: The special case was only triggered if the zero value used
hexadecimal format (which is used by Z3). For example, Bitwuzla prints
using binary format, so this special handling was not triggered for
Bitwuzla.