Fix implementation of Ord for Prop
Previously, the implementation of `<=` for Prop would not obey the
antisymmetry law, I believe in the case of `PEq` specifically.
This leads to weird behaviour, namely that duplicates are not properly
recognized in Set.

The proposed change is to implement `compare` directly, which I think is
easier than trying to fix `<=`.