Fix unsat cache in equivalence checking
Recently, in a refactoring, the write to the cache disappeared, so the
cache was actually always empty.
Here we bring back writing to the cache after unsat result.

Moreover, we make additional changes related to the cache:
Given a set of proposition, we check if it is a *superset* of any
known unsatisfiable set of propositions. The previous check for subset
relation was not correct: a subset of an unsatiafiable set can be
satisfiable.
Related to that, we order the queries from the smallest to largest, so
that we increase the likelyhood of cache hits.