Solidity: Refactor reading standard JSON from solc
We refactor the method readStdJSON so that it can better propagate
errors, similar to other JSON reading functions.

To enable that we introduce a datatype to represent the contract data in
the JSON representation. This simplifies JSON parsing significantly,
because we can rely on Aeson to do the heavy lifting, for the most part.

Additionally, we change the return type to only `Contracts`, since this
is the only part that callers were interested in. We do not try to
construct Asts and Sources anymore.

We should probably not export the function at all and keep it internal.