SymExec: Pre-filter safe branches
Not long ago, we were pre-filtering branches that could not lead to
postcondition violation.  This relied on calling simplification and the
constraints being simplified to [PBool False].  However, recently (in
c34d4d69c), the code changed and it no longer simplified the constraints
before sending them further down the pipeline, and thus the statically
safe branches were no longer filtered before we started asking about
their feasibility.

I believe we should bring back the filtering, if only to have an
information about how many branches lead to Failure that we are
interested in.