SMT: Fix collecting storage reads
We are interested in storage reads where the base storage is abstract.
Previously, `findStorageReads` would also try to extract symbolic
address from a store to concrete storage, if the SLoad expression
contained also AbstractStore as a subexpression.
This could happen if the AbstractStore was part of the value being
stored in the concrete store.
This would previously result in internal error.
Now we check if the base storage is abstract, instead of checking if the
abstract store is present somewhere in the whole expression.